+++ lib/src/format/birthday.dart
@@ -2,7 +2,7 @@ import 'package:intl/intl.dart';
 import 'package:hive_dart/src/constant/date.dart';
 
 /// 把时间戳格式化为 10.01 格式
-String formatBirthday(int timestamp) {
-  final date = DateTime.fromMillisecondsSinceEpoch(timestamp);
+String formatBirthday(int value) {
+  final date = DateTime.fromMillisecondsSinceEpoch(value);
   return DateFormat(DATE_MONTH_DATE_DOT).format(date);
 }