+++ lib/src/format/size.dart
@@ -1,4 +1,3 @@
-import '../is/number.dart';
 import '../constant/size.dart';
 import '../util/number.dart';
 
@@ -6,15 +5,15 @@ import '../util/number.dart';
 String formatSize(num value) {
   if (value >= SIZE_GB) {
     final result = divideNumber(value, SIZE_GB);
-    return '${truncateNumber(result, decimals: isInteger(result) ? 0 : 2)}GB';
+    return '${truncateNumber(result, decimals: hasDecimal(result) ? 2 : 0)}GB';
   }
   else if (value >= SIZE_MB) {
     final result = divideNumber(value, SIZE_MB);
-    return '${truncateNumber(result, decimals: isInteger(result) ? 0 : 2)}MB';
+    return '${truncateNumber(result, decimals: hasDecimal(result) ? 2 : 0)}MB';
   }
   else if (value >= SIZE_KB) {
     final result = divideNumber(value, SIZE_KB);
-    return '${truncateNumber(result, decimals: isInteger(result) ? 0 : 2)}KB';
+    return '${truncateNumber(result, decimals: hasDecimal(result) ? 2 : 0)}KB';
   }
   return '${value}B';
 }