+++ lib/src/is/number.dart
@@ -1,4 +1,0 @@
-/// 是否为整数
-bool isInteger(num value) {
-  return value % 1 == 0;
-}